Move Legacy Code to Submodule
Beginning the 'bait and switch' to the bindgened version of
`onig_sys`.
