+++ onig_sys/src/legacy/constants.rs
